Move NavMenu models from FluxC to app module
The models are only used by the NavMenus feature in the app module,
not by FluxC itself. Moves them to ui/navmenus/models/ for better
code organization.
